vereine notify.* und fensterchef.c
Es ist übersichtlicher, wenn in der Initialisierung von fensterchef
der Graphics Context usw. erstellt wird. Dieser soll später auch für
Input Fenster usw. genutzt werden.
