+++ Makefile
@@ -1,5 +1,5 @@
 # Compiler flags
-DEBUG_FLAGS := -DDEBUG -g -fsanitize=address
+DEBUG_FLAGS := -DDEBUG -DLOG_FILE=\"/tmp/fensterchef-log.txt\" -g -fsanitize=address
 C_FLAGS := -Wall -Wextra -Wpedantic -Werror
 RELEASE_FLAGS := -O3
